refactor: move types where they are used
